refactor template engines
